moved EmailTask class into separate file
